+++ src/Survey/List/Home/Main.tsx
@@ -23,8 +23,7 @@ const HomeMain = ({
   const { url } = useRouteMatch();
   const { navigate } = useContext(NavContext);
 
-  // show activity title.
-  const { activity } = sample.attrs;
+  const { groupId } = sample.attrs;
 
   const isDisabled = sample.isDisabled();
 
@@ -38,9 +37,9 @@ const HomeMain = ({
         )}
 
         {/* Only showing if pre-selected */}
-        {activity && (
+        {groupId && (
           <div className="rounded-list">
-            <MenuAttr.WithLock model={sample} attr="activity" />
+            <MenuAttr.WithLock model={sample} attr="groupId" />
           </div>
         )}